+++ test/host/tests.h
@@ -16,3 +16,4 @@ void run_expedition_defaults_tests(void);
 void run_expedition_logic_tests(void);
 void run_expedition_advance_tests(void);
 void run_hunt_hw_tests(void);
+void run_signal_storm_tests(void);